Add gcore command to gdbinit
This command generates a core dump. The generated core dump can be used
with CrashDebug (https://github.com/adamgreen/CrashDebug) to reload
into GDB at a later point in time or on another machine. The dump will
be generated to a file named "crash.dump"

NOTE: I also renamed the existing "dumpfault" command to "showfault" so
      that it doesn't get confused with what the new gcore command
      does.