implement Module::in_command_ctx() call which is similar to the old on_main_loop() and gets called on demand when in the command thread context where all gcode commands must be executed.
A module can tell it to get the in_command_ctx() callback by setting the module variable Module::want_command_ctx to true.
it can be set to false to not get it. So you can do one shot callbacks as required by clearing it in the in_command_ctx() callback.

Make Switch use this new callback to execute output commands as needed.